Create ParticipantChip
